Open/close out_m2 correctly in parallel_to_m2.py
Currently out_m2 file is opened and not closed, which might lead to weird bugs. Opening with *with* should handle everything.

A side note:
I would like to wrap the main functions in parallel_to_m2.py and compare_m2.py in a separate function, which is then called from the main(): in this way they can be run both as a script (as currently suggested in the readme), or loaded as a function from an external python script. Would you be okay with such a pull request?